requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.4
colorama==0.4.4
comtypes==1.1.10
cryptography==36.0.1
dnspython==2.2.1
docopt==0.6.2
email-validator==1.2.1
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.5.1
greenlet==1.1.2
gspread==5.3.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.1
jaraco.context==4.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
more-itertools==8.12.0
MouseInfo==0.1.3
newsapi-python==0.2.6
num2words==0.5.10
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.1
pdfminer.six==20211012
pdfplumber==0.6.0
pefile==2021.9.3
Pillow==9.0.1
psutil==5.9.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
pyjokes==0.6.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pywin32==303
pywin32-ctypes==0.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.32
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
Wand==0.6.7
Werkzeug==2.0.3
wikipedia==1.4.0
wolframalpha==5.0.0
wsproto==1.0.0
WTForms==3.0.1
xmltodict==0.12.0
